float => double in costs

+++ database/migrations/2018_12_03_152649_create_components_table.php
@@ -19,7 +19,7 @@ class CreateComponentsTable extends Migration
             $table->string('title');
             $table->unsignedInteger('count');
             $table->unsignedInteger('vendor_id')->index();
-            $table->float('cost');
+            $table->double('cost');
             $table->timestamps();
             $table->softDeletes();
         });
